+++ src/misc/str.hpp
@@ -93,14 +93,14 @@ struct CStrA : public CBlobT<char>
 	/** Add formated string (like vsprintf) at the end of existing contents. */
 	int AddFormatL(const char *format, va_list args)
 	{
-		uint addSize = max<uint>(strlen(format), 16);
+		size_t addSize = max<size_t>(strlen(format), 16);
 		addSize += addSize / 2;
 		int ret;
 		int err = 0;
 		for (;;) {
 			char *buf = MakeFreeSpace(addSize);
 			ret = vsnprintf(buf, base::GetReserve(), format, args);
-			if (ret >= (int)base::GetReserve()) {
+			if (ret >= base::GetReserve()) {
 				/* Greater return than given count means needed buffer size. */
 				addSize = ret + 1;
 				continue;
